# Conflicts:
#	src/Avalonia.Controls/Flyouts/FlyoutBase.cs
#	src/Avalonia.Controls/Primitives/Popup.cs

+++ src/Avalonia.Controls/NumericUpDown/NumericUpDownValueChangedEventArgs.cs
@@ -4,13 +4,13 @@ namespace Avalonia.Controls
 {
     public class NumericUpDownValueChangedEventArgs : RoutedEventArgs
     {
-        public NumericUpDownValueChangedEventArgs(RoutedEvent routedEvent, double oldValue,  double newValue) : base(routedEvent)
+        public NumericUpDownValueChangedEventArgs(RoutedEvent routedEvent, decimal oldValue, decimal newValue) : base(routedEvent)
         {
             OldValue = oldValue;
             NewValue = newValue;
         }
 
-        public double OldValue { get; }
-        public double NewValue { get; }
+        public decimal OldValue { get; }
+        public decimal NewValue { get; }
     }
 }
